Fix currencies displaying properly.

+++ lib/api/Currency.dart
@@ -15,7 +15,7 @@ class Currency{
     );
   }
 
-  num rate(Currency to, {int amount = 1}){
+  num rate(Currency to, {double amount = 1}){
   if(this.base != to.base)
   throw Exception("not same base in currencies:" + this.symbol+"-" + this.base +", " + to.symbol + "-" + to.base);
   return amount * this.price/to.price;
